finished

+++ tools/db_handler.py
@@ -6,7 +6,7 @@ import os
 import openpyxl
 import psycopg2
 from loguru import logger
-from tools.sql_commands import COMMANDS, create
+from tools.sql_commands import COMMANDS, create, insert, drop
 
 
 class DBHandler:
@@ -14,7 +14,7 @@ class DBHandler:
 
     """
 
-    def __init__(self, config, data_file, args=None):
+    def __init__(self, config, data_file=None, args=None):
         self.conn = None
         self.cur = None
         self.columns = None
@@ -23,7 +23,6 @@ class DBHandler:
             default_data = json.load(config_file)
             if args is not None:
                 args_dict = vars(args)
-                print(args)
                 for key, val in args_dict.items():
                     if not val:
                         args_dict[key] = default_data[key]
@@ -41,6 +40,9 @@ class DBHandler:
             list: data from .xlsx in python's list of lists
 
         """
+        if not self.filename:
+            logger.error('No datafile')
+            return None
         workbook = openpyxl.load_workbook(self.filename)
         worksheet = workbook.active
         table = [[] for i in range(worksheet.max_row)]
@@ -49,7 +51,7 @@ class DBHandler:
                 if cell.value and cell.value != '':
                     table[row].append(cell.value)
                 else:
-                    table[row].append(None)
+                    table[row].append('None')
 
         for i in table:
             if not i:
@@ -76,14 +78,14 @@ class DBHandler:
             self.cur = self.conn.cursor()
             logger.info("Success!")
         except psycopg2.DatabaseError as error:
-            logger.info(error)
+            logger.error(error)
 
     def drop_table(self):
         logger.info("Creating table....")
         if not self.conn:
             self.connect()
         try:
-            self.cur.execute(COMMANDS["drop"])
+            self.cur.execute(drop(self.config_data["table"]))
             self.conn.commit()
             logger.info("Success!")
         except psycopg2.DatabaseError as error:
@@ -98,7 +100,7 @@ class DBHandler:
         if not self.conn:
             self.connect()
         try:
-            self.cur.execute(create(self.columns))
+            self.cur.execute(create(self.config_data["table"].strip('\''), self.columns))
             self.conn.commit()
             logger.info("Success!")
         except psycopg2.DatabaseError as error:
@@ -109,10 +111,8 @@ class DBHandler:
         """Debug function - prints db content
 
         """
-        if not self.conn:
-            self.connect()
         try:
-            self.cur.execute(COMMANDS["show"])
+            self.cur.execute(COMMANDS["show"], self.config_data["table"])
             column_names = [desc[0] for desc in self.cur.description]
             for i in column_names:
                 print("|" + i + "|", end="")
@@ -130,20 +130,41 @@ class DBHandler:
         """Put data from .xlsx to db
 
         """
+        if not self.filename:
+            logger.error('No datafile')
+            exit(1)
         logger.info("Inserting data to DB....")
         if not self.conn:
             self.connect()
         try:
-            self.cur.executemany(COMMANDS["insert"], self.table)
+            self.cur.execute(insert(self.config_data["table"].strip('\''), self.table))
             self.conn.commit()
             logger.info("Success!")
         except psycopg2.DatabaseError as error:
             logger.info(error)
             self.connect()
 
+    def get_tables(self):
+        """Put data from .xlsx to db
+
+                """
+        logger.info("Inserting data to DB....")
+        if not self.conn:
+            self.connect()
+        try:
+            self.cur.execute(COMMANDS["tables"])
+            self.conn.commit()
+            logger.info("Success!")
+            return self.cur.fetchall()
+        except psycopg2.DatabaseError as error:
+            logger.info(error)
+            self.connect()
+
     def run(self):
+        if not self.filename:
+            logger.error('No datafile')
+            exit(1)
         if self.config_data["overwrite"]:
             self.drop_table()
         self.create_table()
         self.fill_table()
-
